added distance changes

+++ app/src/main/java/enseirb/fr/therunner/RunController.java
@@ -5,16 +5,16 @@ import java.util.Date;
 public class RunController {
     private int id;
     private long time;
-    private int distance;
+    private double distance;
     private int runningUser;
 
-    public RunController(int id, long time, int distance, int user){
+    public RunController(int id, long time, double distance, int user){
         this.id = id;
         this.time = time;
         this.distance = distance;
         runningUser = user;
     }
-    public RunController(long time, int distance, int user){
+    public RunController(long time, double distance, int user){
         this.time = time;
         this.distance = distance;
         this.runningUser = user;
@@ -28,7 +28,7 @@ public class RunController {
         return time;
     }
 
-    public int getDistance(){
+    public double getDistance(){
         return distance;
     }
